% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkjava}
\alias{checkjava}
\title{Check if Java is running}
\usage{
checkjava()
}
\value{
An R logical value (\code{TRUE} or \code{FALSE})
}
\description{
Check if Java is running to determine if GSEA should be performed.
}
\examples{
checkjava()
}
