% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rtocpbool}
\alias{rtocpbool}
\title{Convert R Logical to Command Prompt Logical}
\usage{
rtocpbool(boolIn)
}
\arguments{
\item{boolIn}{A logical value in R (\code{TRUE} or \code{FALSE}).}
}
\value{
A logical value in Command Prompt (\code{true} or \code{false}).
}
\description{
Converts an R logical value (\code{TRUE} or \code{FALSE}) to a format compatible with command prompt.
}
\examples{
rtocpbool(TRUE)
rtocpbool(FALSE)
}
