% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rtocppath}
\alias{rtocppath}
\title{Convert Filename from R Format to Command Prompt Format}
\usage{
rtocppath(strIn)
}
\arguments{
\item{strIn}{A character string representing a file path.}
}
\value{
A Command Prompt path string.
}
\description{
Converts an R string representing a file path to a format compatible with Command Prompt handling.
}
\examples{
rtocppath("C:/Users/Example/Documents/file.txt")
}
