% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{runGSEAPrerank}
\alias{runGSEAPrerank}
\title{Run Preranked Gene Set Enrichment Analysis (GSEAPrerank) via Command Prompt}
\usage{
runGSEAPrerank(
  rnkFile,
  GSEArepLbl,
  GSEAresOutDir,
  gmtFile,
  pathToBat = NA,
  altDelim = NA,
  ChipFile = NA,
  gmtAbbr = NA,
  scoring_scheme = "weighted",
  collapse = "Collapse",
  modeProbe = "Max_probe",
  normMode = "meandiv",
  nperm = 1000,
  create_svgs = FALSE,
  include_only_symbols = TRUE,
  make_sets = TRUE,
  zip_report = FALSE,
  numMarkers = 100,
  plot_top_x = 25,
  rnd_seed = "timestamp",
  set_max = 500,
  set_min = 15,
  showOutputOnConsole = TRUE
)
}
\arguments{
\item{rnkFile}{Path to the RNK file containing the ranked gene list.}

\item{GSEArepLbl}{Label for the GSEA report.}

\item{GSEAresOutDir}{Directory to save GSEA results.}

\item{gmtFile}{Path to the GMT file containing gene sets.}

\item{pathToBat}{Optional path to the GSEA batch file.}

\item{altDelim}{Optional alternate delimiter for the RNK file.}

\item{ChipFile}{Optional path to the chip annotation file.}

\item{gmtAbbr}{Optional abbreviation for the GMT file.}

\item{scoring_scheme}{Scoring scheme to use. Choices: \code{"weighted"}, \code{"classic"}, \code{"weighted_p2"},
\code{"weighted_p1.5"}.}

\item{collapse}{Collapse mode. Choices: \code{"Collapse"}, \code{"No_Collapse"}, \code{"Remap_Only"}.}

\item{modeProbe}{Probe mode. Choices: \code{"Max_probe"}, \code{"Median_of_probe"}, \code{"Mean_of_probe"},
\code{"Sum_of_probes"}, \code{"Abs_max_of_probes"}.}

\item{normMode}{Normalization mode. Choices: \code{"meandiv"}, \code{"none"}.}

\item{nperm}{Number of permutations to perform.}

\item{create_svgs}{Logical; whether to create SVG plots.}

\item{include_only_symbols}{Logical; whether to include only gene symbols.}

\item{make_sets}{Logical; whether to generate gene sets.}

\item{zip_report}{Logical; whether to zip the GSEA report.}

\item{numMarkers}{Number of top markers to include.}

\item{plot_top_x}{Number of top gene sets to plot.}

\item{rnd_seed}{Random seed. Choices: \code{"timestamp"}, \code{"149"}.}

\item{set_max}{Maximum gene set size.}

\item{set_min}{Minimum gene set size.}

\item{showOutputOnConsole}{Logical; whether to show output on console.}
}
\value{
No return value. Results are saved to the specified output directory.
}
\description{
This function performs a preranked Gene Set Enrichment Analysis (GSEA) using a ranked gene list.
It supports various customization options including scoring schemes, probe handling, normalization,
and output formatting. Defaults are set to current GSEA defaults
}
\examples{
runGSEAPrerank("ranked_genes.rnk", "GSEA_Prerank_Rep1", "results/",
               "genesets.gmt")
}
