% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cls.R
\name{write_cls}
\alias{write_cls}
\title{Create .cls File}
\usage{
write_cls(annotDF, filenameOut, groupingColName = 1)
}
\arguments{
\item{annotDF}{A dataframe object with corresponding sample information and a column with sample groups}

\item{filenameOut}{Destination of file}

\item{groupingColName}{The column with group names for the samples}
}
\value{
silently returns NULL
}
\description{
Creates a .cls file for running GSEA
}
\details{
The sample annotation dataframe with a column - default to first - that indicates the group a sample belongs.
}
