% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gct.R
\name{write_gct}
\alias{write_gct}
\title{Proteomics Toolset for Integrative Data Analysis (ProTIGY)}
\usage{
write_gct(ds, ofile, precision = 4, appenddim = TRUE, ver = 3)
}
\arguments{
\item{ds}{the GCT object}

\item{ofile}{the desired output filename}

\item{precision}{the numeric precision at which to
save the matrix. See \code{details}.}

\item{appenddim}{boolean indicating whether to append
matrix dimensions to filename}

\item{ver}{the GCT version to write. See \code{details}.}
}
\value{
silently returns NULL
}
\description{
Copyright ? 2015, The Broad Institute, Inc. (All rights reserved).
}
\details{
Redistribution and use in source and binary forms, with or without modification, are permitted provided that
the following conditions are met:

Redistributions of source code must retain the above copyright notice, this list of conditions and the following
disclaimer.

Redistributions in binary form must reproduce the above copyright notice, this list of conditions
and the following disclaimer in the documentation and/or other materials provided with the distribution.

Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
products derived from this software without specific prior written permission.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
Write a GCT object to disk in GCT format

Since GCT is text format, the higher \code{precision}
you choose, the larger the file size.
\code{ver} is assumed to be 3, aka GCT version 1.3, which supports
embedded row and column metadata in the GCT file. Any other value
passed to \code{ver} will result in a GCT version 1.2 file which
contains only the matrix data and no annotations.
}
\examples{
# note this will create a GCT file in your current directory
write_gct(ds, "dataset", precision=2)

}
\seealso{
Other GCTX parsing functions: 
\code{\link{GCT}}
}
\concept{GCTX parsing functions}
